# docs/design/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dotick Identity API",
    "version": "0.2.0",
    "description": "Executable Increment 1 contract for verified email identity and revocable JWT sessions."
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
...
    "/api/v1/auth/email/resend": {
      "post": {
        "operationId": "resendEmailVerification",
        "tags": ["Identity"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/EmailRequest" } }
          }
        },
        "responses": {
          "202": {
            "description": "Request accepted without disclosing account existence.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Accepted" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/auth/email/verify": {
      "post": {
        "operationId": "verifyEmail",
        "tags": ["Identity"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerificationRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Email verified." },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input or verification code.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
...
    },
    "schemas": {
      "Accepted": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status"],
        "properties": { "status": { "const": "accepted" } }
      },
      "EmailRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email"],
        "properties": { "email": { "type": "string", "format": "email", "maxLength": 254 } }
      },
...
      "VerificationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "code"],
        "properties": {
          "email": { "type": "string", "format": "email", "maxLength": 254 },
          "code": { "type": "string", "pattern": "^[0-9]{6}$" }
        }
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code"],
            "properties": {
              "code": {
                "type": "string",
                "examples": [
                  "validation_error",
                  "invalid_verification_code",
                  "authentication_failed",
                  "token_not_valid",
                  "not_authenticated",
                  "not_found"
                ]
              },
              "details": {}
            }
          }
        }
      }
    }
  }
}